% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_events.R
\name{get_bpm}
\alias{get_bpm}
\title{Get beats per minute}
\usage{
get_bpm(df_meta)
}
\arguments{
\item{df_meta}{data.frame containing midi meta event data resulting of miditapyr$split_df()}
}
\value{
integer of beats per minute
}
\description{
Get beats per minute
}
\examples{
## test
mt <- reticulate::import("miditapyr")
mido <- reticulate::import("mido")
mid_file_str <- system.file("extdata", "test_midi_file.mid", package = "pyramidi")
mido_mid_file <- mido$MidiFile(mid_file_str)
dfc <- mt$midi_to_df(mido_mid_file)
ticks_per_beat = mido_mid_file$ticks_per_beat
df <- dfc \%>\%
  mt$tidy_df()
dfm <- tab_measures(df, ticks_per_beat)
library(zeallot)
mt$split_df(dfm)  \%->\% c(df_meta, df_notes)
get_bpm(df_meta)
}
