% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_events.R
\name{get_tempo}
\alias{get_tempo}
\title{Get tempo}
\usage{
get_tempo(df_meta)
}
\arguments{
\item{df_meta}{data.frame containing midi meta event data resulting of miditapyr$split_df()}
}
\value{
tempo integer
}
\description{
Get tempo
}
\examples{
\donttest{
mid_file_str <- system.file("extdata", "test_midi_file.mid", package = "pyramidi")
mido_mid_file <- mido$MidiFile(mid_file_str)
dfc <- miditapyr$midi_to_df(mido_mid_file)
ticks_per_beat = mido_mid_file$ticks_per_beat
df <- dfc \%>\%
  miditapyr$tidy_df()
}
\dontrun{
dfm <- tab_measures(df, ticks_per_beat)
library(zeallot)
miditapyr$split_df(dfm)  \%->\% c(df_meta, df_notes)
get_tempo(df_meta)
}
}
