% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_python_pkgs.R
\name{install_miditapyr}
\alias{install_miditapyr}
\title{Install miditapyr python package}
\usage{
install_miditapyr(envname = NULL, method = "auto", conda = "auto", pip = TRUE)
}
\arguments{
\item{envname}{The name, or full path, of the environment in which python
packages are to be installed. When `NULL` (default), the active environment
(`RETICULATE_PYTHON_ENV` variable) will be used; if that is unset, then the
`"r-reticulate"` environment will be used.}

\item{method}{Installation method. By default, "auto" automatically finds a
method that will work in the local environment. Change the default to force
a specific installation method. Note that the "virtualenv" method is not
available on Windows.}

\item{conda}{The path to a conda executable. Use "auto" to allow reticulate
to automatically find an appropriate conda binary. See Finding Conda for
more details.}

\item{pip}{Boolean; use pip for package installation? This is only relevant
when Conda environments are used, as otherwise packages will be installed
from the Conda repositories.}
}
\description{
Wrapper around `reticulate::py_install()` to install the miditapyr python package
}
\details{
From `reticulate::py_install()`: On Linux and OS X the "virtualenv"
  method will be used by default ("conda" will be used if virtualenv isn't
  available). On Windows, the "conda" method is always used.
}
