% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pivot_note_on_off.R
\name{merge_long_events}
\alias{merge_long_events}
\title{Merge dataframes transformed back to long format}
\usage{
merge_long_events(df_meta, df_notes_long, df_not_notes)
}
\arguments{
\item{df_meta, df_notes_long, df_not_notes}{Results of \code{miditapyr$split_df()}.}
}
\value{
Merges the input dataframes, arranges by \code{i_track} & (absolute) \code{ticks}, and
calculates \code{time} (in relative ticks since the last event).
}
\description{
Merge dataframes transformed back to long format, remove added columns and
transform to the right chronological order in order to replace the original
midi_frame_unnested object.
}
\examples{

midi_file_string <- system.file("extdata", "test_midi_file.mid", package = "pyramidi")
mf <- miditapyr$MidiFrames(midi_file_string)

dfm <- tab_measures(mf$midi_frame_unnested$df, ticks_per_beat = mf$midi_file$ticks_per_beat)

l <- split_midi_frame(dfm)

df_notes_long <- pivot_long_notes(l$df_notes_wide)

merge_long_events(l$df_meta, df_notes_long, l$df_not_notes)
}
\seealso{
Other Split/merge meta/notes/non-note events: 
\code{\link{split_midi_frame}()}
}
\concept{Split/merge meta/notes/non-note events}
