% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{midi_defs}
\alias{midi_defs}
\title{Table of midi notes}
\format{
A data frame with 127 rows and 2 variables:
\describe{
  \item{note}{integer from 0 - 127 containing the midi pitches}
  \item{note_name}{name of the note and the octave}
  ...
}
}
\usage{
midi_defs
}
\description{
A table containing the midi pitches and their note names
}
\keyword{datasets}
