% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/python_pkgs.R
\docType{data}
\name{miditapyr}
\alias{miditapyr}
\title{Miditapyr python module}
\format{
An object of class \code{python.builtin.module} (inherits from \code{python.builtin.object}) of length 1.
}
\usage{
miditapyr
}
\value{
reticulate python module
}
\description{
Miditapyr python module
}
\examples{
\dontrun{
midi_file_string <- system.file("extdata", "test_midi_file.mid", package = "pyramidi")
miditapyr$MidiFile(midi_file_string)
}
}
\keyword{datasets}
