% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_python_pkgs.R
\docType{data}
\name{mido}
\alias{mido}
\title{Mido python module}
\format{
An object of class \code{python.builtin.module} (inherits from \code{python.builtin.object}) of length 2.
}
\usage{
mido
}
\value{
reticulate python module
}
\description{
Mido python module
}
\examples{
\dontrun{
content <- system.file("example_files", "Bass_sample.mid", package="tuneR")
mido$MidiFile(content)
}
}
\keyword{datasets}
