% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/midi_frames.R
\name{new_midi_framer}
\alias{new_midi_framer}
\title{Construct a midi frame object}
\usage{
new_midi_framer(midi_file_string)
}
\arguments{
\item{midi_file_string}{Path to the midi file}
}
\value{
Midi frames S3 object
}
\description{
Constructor of an object of class "midi_framer".
}
\examples{
midi_file_string <- system.file("example_files", "Bass_sample.mid", package="tuneR")
midi_file_string <- system.file("extdata", "test_midi_file.mid", package = "pyramidi")
new_midi_framer(midi_file_string)
}
