% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{piano_keys_coordinates}
\alias{piano_keys_coordinates}
\title{Table of piano keys coordinates}
\format{
A data frame with 127 rows and 6 variables:
\describe{
  \item{midi}{integer from 0 - 127 containing the midi pitches}
  \item{layer}{integer 1 or 2 depending on whether key is white (1) or black (2)}
  \item{xmin}{xmin coordinate of the key}
  \item{ymin}{ymin coordinate of the key}
  \item{xmax}{xmax coordinate of the key}
  \item{ymax}{ymax coordinate of the key}
  ...
}
}
\usage{
piano_keys_coordinates
}
\description{
A table containing the piano keys coordinates. Each piano key is centered at
the midi note value it represents.
}
\examples{
library(ggplot2)
# Print dataframe:
piano_keys_coordinates
# Plot the keyboard:
piano_keys_coordinates \%>\%
  # plot white keys first that they don't cover half of the black keys:
  dplyr::arrange(layer) \%>\%
  ggplot(aes(xmin = -ymin, xmax = -ymax, ymin = xmin, ymax = xmax, fill = factor(layer))) +
  geom_rect(color = "black", show.legend = FALSE) +
  scale_fill_manual(values = c("#ffffdd", "#113300")) +
  theme_void() +
  coord_fixed(ratio = 0.1)

}
\keyword{datasets}
