% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pivot_note_on_off.R
\name{pivot_long_notes}
\alias{pivot_long_notes}
\title{Write note_on and _off events in two lines (wide to long)}
\usage{
pivot_long_notes(df_notes_wide)
}
\arguments{
\item{df_notes_wide}{notes dataframe in wide format}
}
\value{

}
\description{
Write note_on and _off events in two lines (wide to long)
}
\examples{
\dontrun{
dfm <- tab_measures(df, ticks_per_beat)
dfw <- dfm \%>\% pivot_wide_notes()


library(zeallot)
c(df_meta, df_notes) \%<-\% miditapyr$split_df(dfm)
}
}
