% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3midi.R
\name{r_midi_frames}
\alias{r_midi_frames}
\title{Construct a midi frame object}
\usage{
r_midi_frames(midi_file_string)
}
\arguments{
\item{midi_file_string}{Path to the midi file}
}
\value{
Midi frames S3 object
}
\description{
Constructor of an object of class "r_midi_frames".
}
\examples{
midi_file_string <- system.file("extdata", "test_midi_file.mid", package = "pyramidi")
r_midi_frames(midi_file_string)
}
