% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_events.R
\name{tab_time_sig}
\alias{tab_time_sig}
\title{Tabulate time signature}
\usage{
tab_time_sig(df_meta)
}
\arguments{
\item{df_meta}{data.frame containing midi meta event data resulting of miditapyr$mido_midi_df()}
}
\value{
time signature data.frame
}
\description{
Tabulate time signature
}
\examples{
library(zeallot)
mt <- reticulate::import("miditapyr")
mido <- reticulate::import("mido")
mid_file <- system.file("extdata", "test_midi_file.mid", package = "pyramidi")
mido$MidiFile(mid_file) \%>\%
  mt$mido_midi_df()  \%->\% c(df_meta, df_notes, ticks_per_beat)
pyramidi:::tab_time_sig(df_meta)
}
