% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pivot_note_on_off.R
\name{triage_measured_tidy}
\alias{triage_measured_tidy}
\title{Split tidy midi dataframe into parts}
\usage{
triage_measured_tidy(dfm)
}
\arguments{
\item{dfm}{result of \code{tab_measures()}}
}
\value{

}
\description{
Tidy midi frame from `\code{tab_measures()} is split into 3 parts:
df_meta, df_not_notes & df_notes_wide.
}
\examples{
\dontrun{
mid_file_str <- system.file("extdata", "test_midi_file.mid", package = "pyramidi")
mido_mid_file <- mido$MidiFile(mid_file_str)
dfc <- miditapyr$midi_to_df(mido_mid_file)
ticks_per_beat = mido_mid_file$ticks_per_beat
df <- dfc \%>\%
  miditapyr$tidy_df()
}
\dontrun{
dfm <- tab_measures(df, ticks_per_beat)
triage_measured_tidy(dfm)
}
}
