% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/widen.R
\name{widen_events}
\alias{widen_events}
\title{write note_on and _off events in the same line (long to wide)}
\usage{
widen_events(df_measures)
}
\arguments{
\item{df_measures}{}
}
\value{

}
\description{
write note_on and _off events in the same line (long to wide)
}
