% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_events.R
\name{get_bpm}
\alias{get_bpm}
\title{Get beats per minute}
\usage{
get_bpm(df_meta)
}
\arguments{
\item{df_meta}{data.frame containing midi meta event data resulting of miditapyr$mido_midi_df()}
}
\value{
integer of beats per minute
}
\description{
Get beats per minute
}
\examples{
library(zeallot)
mt <- reticulate::import("miditapyr")
mido <- reticulate::import("mido")
mid_file <- system.file("extdata", "test_midi_file.mid", package = "pyramidi")
mido$MidiFile(mid_file) \%>\%
  mt$mido_midi_df()  \%->\% c(df_meta, df_notes, ticks_per_beat)
get_bpm(df_meta)
}
