% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_events.R
\name{get_bpm}
\alias{get_bpm}
\title{Get beats per minute}
\usage{
get_bpm(df_meta)
}
\arguments{
\item{df_meta}{data.frame containing midi meta event data resulting of miditapyr$split_df()}
}
\value{
integer of beats per minute
}
\description{
Get beats per minute
}
\examples{
\dontrun{
dfm <- tab_measures(df, ticks_per_beat)
library(zeallot)
miditapyr$split_df(dfm)  \%->\% c(df_meta, df_notes)
get_bpm(df_meta)
}
}
