% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_events.R
\name{get_tempo}
\alias{get_tempo}
\title{Get tempo}
\usage{
get_tempo(df_meta)
}
\arguments{
\item{df_meta}{data.frame containing midi meta event data resulting of miditapyr$split_df()}
}
\value{
tempo integer
}
\description{
Get tempo
}
\examples{
\dontrun{
dfm <- tab_measures(df, ticks_per_beat)
library(zeallot)
miditapyr$split_df(dfm)  \%->\% c(df_meta, df_notes)
get_tempo(df_meta)
}
}
