% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/python_pkgs.R
\docType{data}
\name{mido}
\alias{mido}
\title{Mido python module}
\format{
An object of class \code{python.builtin.module} (inherits from \code{python.builtin.object}) of length 0.
}
\usage{
mido
}
\value{
reticulate python module
}
\description{
Mido python module
}
\examples{
\dontrun{
midi_file_string <- system.file("extdata", "test_midi_file.mid", package = "pyramidi")
mido$MidiFile(midi_file_string)
}
}
\keyword{datasets}
