% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pivot_note_on_off.R
\name{pivot_long_notes}
\alias{pivot_long_notes}
\title{Write \code{note_on} and \code{note_off} events in two lines (wide to long)}
\usage{
pivot_long_notes(df_notes_wide)
}
\arguments{
\item{df_notes_wide}{notes dataframe in wide format}
}
\value{
Transforms notes in wide dataframe format to long format.
}
\description{
Write \code{note_on} and \code{note_off} events in two lines (wide to long)
}
\examples{
\dontrun{
mid_file_str <- system.file("extdata", "test_midi_file.mid", package = "pyramidi")
mido_mid_file <- mido$MidiFile(mid_file_str)
dfc <- miditapyr$frame_midi(mido_mid_file)
ticks_per_beat = mido_mid_file$ticks_per_beat
df <- dfc \%>\%
  miditapyr$unnest_midi()
}
\dontrun{
dfm <- tab_measures(df, ticks_per_beat)
library(zeallot)
c(df_meta, df_notes) \%<-\% miditapyr$split_df(dfm)
dfw <- df_notes \%>\% pivot_wide_notes()
dfw \%>\% pivot_long_notes()
}
}
\seealso{
Other Pivot midi frame functions: 
\code{\link{pivot_wide_notes}()}
}
\concept{Pivot midi frame functions}
