% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pivot_note_on_off.R
\name{pivot_long_notes}
\alias{pivot_long_notes}
\title{Write note_on and _off events in two lines (wide to long)}
\usage{
pivot_long_notes(df_notes_wide)
}
\arguments{
\item{df_notes_wide}{notes dataframe in wide format}
}
\value{

}
\description{
Write note_on and _off events in two lines (wide to long)
}
\examples{
\dontrun{
mid_file_str <- system.file("extdata", "test_midi_file.mid", package = "pyramidi")
mido_mid_file <- mido$MidiFile(mid_file_str)
dfc <- miditapyr$midi_to_df(mido_mid_file)
ticks_per_beat = mido_mid_file$ticks_per_beat
df <- dfc \%>\%
  miditapyr$tidy_df()
}
\dontrun{
dfm <- tab_measures(df, ticks_per_beat)
dfw <- dfm \%>\% widen_events()


library(zeallot)
c(df_meta, df_notes) \%<-\% miditapyr$split_df(dfm)
}
}
