% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pivot_note_on_off.R
\name{pivot_wide_notes}
\alias{pivot_wide_notes}
\title{Write \code{note_on} and \code{note_off} events in the same line (long to wide)}
\usage{
pivot_wide_notes(df_measures)
}
\arguments{
\item{df_measures}{data.frame resulting of miditapyr$mido_midi_df() and then running tab_measures() (see example)}
}
\value{
A data.frame with the following columns pivoted to wide: c("m", "b", "t", "ticks", "time", "velocity"). Every column is pivoted to wide with the suffix "_note_on" & "_note_off" showing the values of the original column as a prefix. See \code{?tab_measures} for an explanation of the meaning of these columns.
}
\description{
Write \code{note_on} and \code{note_off} events in the same line (long to wide)
}
\examples{
\dontrun{
mid_file_str <- system.file("extdata", "test_midi_file.mid", package = "pyramidi")
mido_mid_file <- mido$MidiFile(mid_file_str)
dfc <- miditapyr$frame_midi(mido_mid_file)
ticks_per_beat = mido_mid_file$ticks_per_beat
df <- dfc \%>\%
  miditapyr$unnest_midi()
}
\dontrun{
dfm <- tab_measures(df, ticks_per_beat)
library(zeallot)
c(df_meta, df_notes) \%<-\% miditapyr$split_df(dfm)
df_notes \%>\% pivot_wide_notes()
}
}
\seealso{
Other Pivot midi frame functions: 
\code{\link{pivot_long_notes}()}
}
\concept{Pivot midi frame functions}
