% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/audio.R
\name{play_midi_frame}
\alias{play_midi_frame}
\title{Play midi from midi frame}
\usage{
play_midi_frame(mfr, audiofile = tempfile("mf_out_", fileext = ".wav"))
}
\arguments{
\item{mfr}{r_midi_frames object}

\item{audiofile}{audiofile}
}
\value{

}
\description{
Transform the midi file to fileext format and provide a play button for html
documents.
}
\examples{
\dontrun{
midi_file_string <- system.file("example_files", "Bass_sample.mid", package="tuneR")
midi_file_string <- system.file("extdata", "test_midi_file.mid", package = "pyramidi")
mfr <- r_midi_frames(midi_file_string)
mfr \%>\% play_midi_frame()
}
}
