% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/audio.R
\name{player}
\alias{player}
\title{Play midi file}
\usage{
player(audiofile, fileext = "wav")
}
\arguments{
\item{audiofile}{audiofile}

\item{fileext}{file extension}
}
\value{
html tag of a playbutton
}
\description{
Transform the midi file to fileext format and provide a play button for html
documents.
}
\examples{
\dontrun{
midifile <- system.file("extdata", "test_midi_file.mid", package = "pyramidi")
audiofile <- "test.wav"
mp3file <- "test.mp3"
fluidsynth::midi_convert(midifile, output = audiofile)
av::av_audio_convert(audiofile, mp3file)
audiofile \%>\% player()
}
}
