% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pivot_note_on_off.R
\name{split_midi_frame}
\alias{split_midi_frame}
\title{Split unnested midi dataframe into parts}
\usage{
split_midi_frame(dfm)
}
\arguments{
\item{dfm}{result of \code{tab_measures()}}
}
\value{
df_meta, df_not_notes & df_notes_wide
}
\description{
Unnested midi frame with time data from \code{tab_measures()} is split into 3 parts:
\itemize{
\item{df_meta: Consisting of all the \href{https://mido.readthedocs.io/en/latest/midi_files.html#meta-messages}{meta} events in the midi data.}
\item{df_not_notes: Containing all midi events that are not \code{meta} and not \code{note_on} or \code{note_off}.}
\item{df_notes_wide: All \code{note_on} or \code{note_off} events and furthermore transformed to wide format by \code{pivot_note_wide}.}
}
}
\examples{
\dontrun{
mid_file_str <- system.file("extdata", "test_midi_file.mid", package = "pyramidi")
mido_mid_file <- mido$MidiFile(mid_file_str)
dfc <- miditapyr$frame_midi(mido_mid_file)
ticks_per_beat = mido_mid_file$ticks_per_beat
df <- dfc \%>\%
  miditapyr$unnest_midi()
}
\dontrun{
dfm <- tab_measures(df, ticks_per_beat)
split_midi_frame(dfm)
}
}
\seealso{
Other Split/merge meta/notes/non-note events: 
\code{\link{merge_midi_frames}()}
}
\concept{Split/merge meta/notes/non-note events}
