% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures.R
\name{tab_measures}
\alias{tab_measures}
\title{Tabulate measure related data in the mdi event data}
\usage{
tab_measures(df_meta, df_notes, ticks_per_beat)
}
\arguments{
\item{df_notes}{}
}
\value{

}
\description{
Tabulate measure related data in the mdi event data
}
\examples{
library(zeallot)
mt <- reticulate::import("miditapyr")
mido <- reticulate::import("mido")
mid_file <- system.file("extdata", "test_midi_file.mid", package = "pyramidi")
mido$MidiFile(mid_file) \%>\%
  mt$mido_midi_df()  \%->\% c(df_meta, df_notes, ticks_per_beat)
  tab_measures(df_meta, df_notes, ticks_per_beat)
}
