% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures.R
\name{tab_measures}
\alias{tab_measures}
\title{Tabulate measure related data in the midi event data}
\usage{
tab_measures(df, ticks_per_beat)
}
\arguments{
\item{df}{Note event data.frame resulting of miditapyr$unnest_midi()}

\item{ticks_per_beat}{integer resulting of miditapyr$mido_midi_df()}
}
\value{
A data.frame with the following columns added: c("m", "b", "t", "ticks", "time"). \cr
  m: measure \cr
  b: beat \cr
  t: time in seconds \cr
}
\description{
Tabulate measure related data in the midi event data
}
\examples{
\dontrun{
tab_measures(df, ticks_per_beat)
}
}
