% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_events.R
\name{tab_time_sig}
\alias{tab_time_sig}
\title{Tabulate time signature}
\usage{
tab_time_sig(df_meta)
}
\arguments{
\item{df_meta}{data.frame containing midi meta event data resulting of miditapyr$split_df()}
}
\value{
time signature data.frame
}
\description{
Tabulate time signature
}
\examples{
\dontrun{
dfm <- tab_measures(df, ticks_per_beat)
library(zeallot)
miditapyr$split_df(dfm)  \%->\% c(df_meta, df_notes)
pyramidi:::tab_time_sig(df_meta)
}
}
