% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pivot_note_on_off.R
\name{triage_measured_unnested}
\alias{triage_measured_unnested}
\title{Split unnested midi dataframe into parts}
\usage{
triage_measured_unnested(dfm)
}
\arguments{
\item{dfm}{result of \code{tab_measures()}}
}
\value{

}
\description{
Unnested midi frame with time data from `\code{tab_measures()} is split into 3 parts:
df_meta, df_not_notes & df_notes_wide.
}
\examples{
\dontrun{
dfm <- tab_measures(df, ticks_per_beat)
triage_measured_unnested(dfm)
}
}
