% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/widen.R
\name{widen_events}
\alias{widen_events}
\title{write note_on and _off events in the same line (long to wide)}
\usage{
widen_events(df_measures)
}
\arguments{
\item{df_measures}{data.frame resulting of miditapyr$mido_midi_df() and then running tab_measures() (see example)}
}
\value{
A data.frame with the following columns pivoted to wide: c("m", "b", "t", "ticks", "time", "velocity"). Every column is pivoted to wide with the suffix "_note_on" & "_note_off" showing the values of the original column as a prefix. See \code{?tab_measures} for an explanation of the meaning of these columns.
}
\description{
write note_on and _off events in the same line (long to wide)
}
\examples{
library(zeallot)
mt <- reticulate::import("miditapyr")
mido <- reticulate::import("mido")
mid_file <- system.file("extdata", "test_midi_file.mid", package = "pyramidi")
mido$MidiFile(mid_file) \%>\%
  mt$mido_midi_df()  \%->\% c(df_meta, df_notes, ticks_per_beat)
  tab_measures(df_meta, df_notes, ticks_per_beat) \%>\%
  widen_events()
}
