% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/widen.R
\name{widen_events}
\alias{widen_events}
\title{write note_on and _off events in the same line (long to wide)}
\usage{
widen_events(df_measures)
}
\arguments{
\item{df_measures}{}
}
\value{

}
\description{
write note_on and _off events in the same line (long to wide)
}
\examples{
library(zeallot)
mid_file <- '/home/chief/Bitwig Studio/Projects/bossa_raw/midi_arpeggiated.mid'
mt <- reticulate::import("miditapyr")
mido <- reticulate::import("mido")
mido$MidiFile(mid_file) \%>\%
  mt$mido_midi_df()  \%->\% c(df_meta, df_notes, ticks_per_beat)
  tab_measures(df_meta, df_notes, ticks_per_beat) \%>\%
  widen_events()
}
